Separate demo of PatchDriver-tests from regression
The test for counter driver is meant to be a flagship example of
PatchDriver testing, for plugin developers to use.

Compared with the (new) Activity test, which is a regressions test for
running a migration then a forge in one go.

Having different purposes, these are meant to be in different location.
Fixed by moving the forge-related tests to separate "activity" tests,
away from counterdriver test folder.