Separate docs for available plugins
Refactor the docs to use a generic plugin-discovery template (with
variation for scanners=func vs the rest=class).

Refactor the discovery of scanners to fit in the rest of the plugins, to
enable that docs discovery.

Change scanner discovery to avoid silent skipping, but now ImportError

Change the index.rst to index.md, as I learned more about myst-parser.